pyproject.toml
[project]
name = "hexa-tic-tac-toe"
version = "0.1.0"
description = "Hexagonal Tic Tac Toe Game"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruff>=0.15.7",
    "ty>=0.0.24",
    "pytest>=8.0.0",
    "pettingzoo>=1.25.0",
    "numpy>=2.4.3",
    "pgx>=2.6.0",
    "flax>=0.12.6",
    "optax>=0.2.8",
    "mctx>=0.0.6",
    "chex>=0.1.91",
    "orbax-checkpoint>=0.11.33",
    "wandb>=0.25.1",
    "flashbax>=0.1.3",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[project.scripts]
hexa = "hexa_tic_tac_toe.cli:main"
hexa-train = "hexa_tic_tac_toe.tools.train:main"
hexa-export = "hexa_tic_tac_toe.tools.export_game:main"
